\name{build_simpop}
\alias{build_simpop}
\title{Build a simulated population by bootstrapping from a list}

\usage{build_simpop(datalist, rowlist, sim)}
\arguments{
  \item{datalist}{List of data frames containing the data to be 
bootstrapped then combined into a single data frame}
  \item{rowlist}{List of data frames containing row numbers for 
bootstrapping the data in datalist}
  \item{sim}{Simulation number. Indicates which column of rowlist
to use}
}



\author{Jeanette Birnbaum & Leslie Mallinger}





