\name{head2dim}
\alias{head2dim}
\title{Display the first six rows and first six columns of a two-dimensional data frame or matrix}

\usage{head2dim(x)}
\arguments{
  \item{x}{A two-dimensional data frame or matrix}
}

\value{A two-dimensional data frame or matrix}

\author{Jeanette Birnbaum & Leslie Mallinger}





