\name{random_rows}
\alias{random_rows}
\title{Randomly samples rows from a data frame}

\usage{random_rows(dset, n, rplc = FALSE)}
\arguments{
  \item{dset}{Data frame from which to sample}
  \item{n}{Number of rows to sample}
  \item{rplc}{Should sampling be with replacement?}
}



\author{Jeanette Birnbaum & Leslie Mallinger}





